Make client input rules amount editable dynamically
